Split part of react jsx rules into new files
There are a few specific ways of writing jsx components, consisting of
multiple rules that together set the way the components, their props and
children are indented, how many props are on the same line, and where
the closing tags and brackets are placed.

This rules were taken out from the main react core rule configuration
and split into two separate configurations, for two different ways of
writing jsx components.

The two new files contain the same rules.

The names of the files might change in the future.